% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMI-package.R
\docType{package}
\name{dMI-package}
\alias{dMI-package}
\title{dMI: dMI: Distance-based Mutual Information Test of Multivariate Association}
\description{
dMI is an R package for estimation of distance-based mutual information and multivariate association testing using a corresponding permutation test. The package implements a nonparametric density estimation procedure of distance-based mutual information using k-nearest neighbor (kNN) distances and a permutation test for independence.
}
\author{
\strong{Maintainer}: Rebecca A. Deek \email{rdeek@pitt.edu}

Authors:
\itemize{
  \item Rebecca A. Deek \email{rdeek@pitt.edu}
}

}
\keyword{internal}
