% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dMI.R
\name{dMI}
\alias{dMI}
\title{Distance-based mutual information}
\usage{
dMI(D.x, D.y, kbar, n.perm = 999, n.cores = 1, MI = FALSE)
}
\arguments{
\item{D.x}{An \eqn{n \times n} symmetric distance matrix of class \code{dist} or
\code{matrix} for modality \eqn{X}. \eqn{n} refers to the total number of observations
(samples) in the data set. Individuals must have observations in both \eqn{X}
and \eqn{Y}.}

\item{D.y}{An \eqn{n \times n} symmetric distance matrix of class \code{dist} or
\code{matrix} for modality \eqn{Y}. \eqn{n} refers to the total number of observations
(samples) in the data set.}

\item{kbar}{A vector of nearest neighbors to be used in density estimation.
It is recommend to use \eqn{1:k}, where \eqn{k} greater than or equal to
\eqn{\sqrt{n(n-1)/2}}. \eqn{k} cannot be larger than \eqn{n(n-1)/2}.}

\item{n.perm}{Number of permutations to be used in hypothesis testing. Default
is set as 999.}

\item{n.cores}{Number of cores to be used in parallel processing. Default is
set to 1 (no parallelization).}

\item{MI}{Logical (\code{TRUE}/\code{FALSE}) indicating if the mutual information
estimate should be returned. Default is \code{FALSE} and suggested if independence
testing is of primary interest.}
}
\value{
A list containing the permutation test p-value for the global null
hypothesis of independence, by default. If \code{MI = TRUE}, the list also contains
the KL estimate of distance-based mutual information.
}
\description{
Function to perform estimation of distance-based mutual information using kNN
density estimation and a permutation test for the global null hypothesis of
independence.
}
\examples{
\dontshow{if (requireNamespace("vegan", quietly = TRUE)) withAutoprint(\{ # examplesIf}
mb.bray = vegan::vegdist(microbiomeData, method = "bray")
host.euclidean = vegan::vegdist(hostData, method = "euclidean")
dMI(D.x = mb.bray, D.y = host.euclidean, kbar = 1:71)
\dontshow{\}) # examplesIf}
}
