% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gowerCentering.R
\name{gower.centering}
\alias{gower.centering}
\title{Gower's centering}
\usage{
gower.centering(D)
}
\arguments{
\item{D}{An \eqn{n \times n} symmetric distance matrix of class \code{dist} or \code{matrix}.
\eqn{n} refers to the total number of observations, or samples, in the data set.}
}
\value{
An \eqn{n \times n} symmetric double centered similarity matrix.
}
\description{
Function to convert a distance matrix into a double centered similarity matrix
using Gower's centering: \eqn{G = -\frac{1}{2} (I - \frac{11^T}{n}) D^2 (I - \frac{11^T}{n})}.
}
\examples{
\dontshow{if (requireNamespace("vegan", quietly = TRUE)) withAutoprint(\{ # examplesIf}
mb.bray = vegan::vegdist(microbiomeData, method = "bray")
D.x = gower.centering(mb.bray)
\dontshow{\}) # examplesIf}
}
