% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hostData.R
\docType{data}
\name{hostData}
\alias{hostData}
\title{Simulated host data}
\format{
An \eqn{n \times q} object of class \code{matrix} contain host omics data
(e.g., metabolite or protein concentrations).
}
\source{
Generated for package examples.
}
\usage{
hostData
}
\description{
Simulated host omics data using a log-normal distribution.
}
\examples{
data(hostData)
str(hostData)
}
\keyword{datasets}
