% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbiomeData.R
\docType{data}
\name{microbiomeData}
\alias{microbiomeData}
\title{Simulated microbiome data}
\format{
An \eqn{n \times p} object of class \code{matrix} contain microbial
sequencing counts.
}
\source{
Generated for package examples.
}
\usage{
microbiomeData
}
\description{
Simulated microbiome data using the \code{sparseDOSSA2} package.
}
\examples{
data(microbiomeData)
str(microbiomeData)
}
\keyword{datasets}
